新增LD文件解析功能代码修改

+++ RESTFUL/src/main/java/com/prs/hub/utils/MultipartFileToFileUtil.java
@@ -68,15 +68,27 @@ public class MultipartFileToFileUtil {
      *
      * @param file
      */
-    public static Boolean delteTempFile(File file) {
-        Boolean flag = false;
-
-        if (file != null) {
-            flag = file.delete();
+    public static void delteTempFile(File file) {
+        if(file != null && file.exists()) {//判断路径是否存在
+            if(file.isFile()){//boolean isFile():测试此抽象路径名表示的文件是否是一个标准文件。
+                file.delete();
+            }else{//不是文件，对于文件夹的操作
+                //保存 路径下的所有的文件和文件夹到listFiles数组中
+                File[] listFiles = file.listFiles();//listFiles方法：返回file路径下所有文件和文件夹的绝对路径
+                for (File file2 : listFiles) {
+                    /*
+                     * 递归作用：由外到内先一层一层删除里面的文件 再从最内层 反过来删除文件夹
+                     *    注意：此时的文件夹在上一步的操作之后，里面的文件内容已全部删除
+                     *         所以每一层的文件夹都是空的  ==》最后就可以直接删除了
+                     */
+                    delteTempFile(file2);
+                }
+            }
+            //删除空文件夹
+            file.delete();
+        }else {
+            log.info("该file路径不存在！！");
         }
-        log.info("删除本地临时文件flag="+flag);
-
-        return flag;
     }
 }
